feat: create useRandomImg hooks, refactor project code
