+++ src/components/RandomImgBox/RandomImgBox.style.ts
